game: clear the last conformance errors outside the vehicle model
Six unrelated MSVC-isms, none of them physics:

- lsl::ResourceCollection declared its Value typedef private, so
  SoundLib could not name it when overriding RemoveItem. Now protected,
  which is what a type used in a virtual's signature has to be.
- GameMode's Node::operator< was non-const, and std::list::sort compares
  through const references.
- Two D3DXVECTOR4-to-D3DXVECTOR2 assignments needing two user-defined
  conversions, and two std::min calls mixing size_t with unsigned.

XPlatform gains AdjustWindowRect, GetCurrentThread and
SetThreadAffinityMask. The affinity call is worth a note: it exists to pin
the main thread to CPU 0 so QueryPerformanceCounter does not jump between
cores on older AMD parts. That is a Windows-specific hazard --
mach_absolute_time is coherent across cores -- so the substitute reports
success and does nothing, which is the correct behaviour rather than a
gap.

Rock3dGame is now down to 20 errors, every one of them in GameCar.cpp and
GameCar.h: the PhysX 2.8 wheel runtime API. 11 of 52 translation units
compile; the rest are blocked behind that single header.

+++ src/Rock3dGame/source/game/MainMenu2.cpp
@@ -1545,7 +1545,7 @@ void NetworkFrame::OnShow(bool value)
 		{
 			lsl::string text = "";
 
-			for (unsigned i = 0; i < std::min(addrVec.size(), 6U); ++i)
+			for (size_t i = 0; i < std::min(addrVec.size(), static_cast<size_t>(6)); ++i)
 				text += addrVec[i] + "\n";
 
 			_labels[mlIPAdress]->SetText(!text.empty() ? "My IP:\n" + text : "");
